Create README.md
This project analyzes the movie landscape curated by Microsoft for its emerging movie studio. A descriptive analysis of the popularity of movies, the average duration of movies, 
and movies with the highest revenue shows that comedy and drama are the most popular genres of Movies, the average duration of a movie is 90 minutes and the genre with the highest 
revenue is drama. The Microsoft company can use this analysis to inform their revenue and budget allocation during the creation of movies and invest most resources in the movies that 
will return the highest revenue.